refactor(ng-update): handle namespaced angular decorators in update-tool
Previously Angular did not support namespaced Angular decorators. e.g.

```
@core.Component({...})
@core.Directive({...})
```

Since this is now supported in ngtsc, we want to handle
such cases where stylesheets, templates are referenced in
namespaced decorators.

This can be achieved by moving the more advanced import
resolution logic made for the Hammer gesture migration
to the `cdk/update-tool`. Exposing these for other CDK
and Material migrations can be helpful anyway.

This also unveiled that we did not run non-misc test cases
automatically. This will be fixed as part of this commit.
