+++ src/material/schematics/ng-update/test-cases/v9/misc/material-imports.spec.ts
@@ -1,5 +1,5 @@
 import {createTestCaseSetup, readFileContent} from '@angular/cdk/schematics/testing';
-import {migrationCollection} from '../index.spec';
+import {migrationCollection} from '../../index.spec';
 
 describe('v9 material imports', () => {
   it('should re-map top-level material imports to the proper entry points when top-level ' +
